adapt for tested SIFT

+++ inc/qbuilder.h
@@ -19,7 +19,7 @@ public:
     vector<bitset<256>> getDescriptor(const string path);
 
 private:
-    bitset<256> p_qlizer(const double *descriptors);
+    bitset<256> p_qlizer(const float *descriptors);
     bool selfInitial;
     Ptr<Feature2D> detector;
 };
